restructuring soundFetch to Main.js function
